chore: moving types into separate file
